Refactor loading bvh files and streaming mocap data
